Fixed Bug

+++ ai_module/extras/swarm_descriptionold.py
@@ -12,43 +12,43 @@ SWARM_DESCRIPTION = {
 },
 {
     "fieldName" : "total_tcp",
-    "fieldType" : "int",
+    "fieldType" : "float",
     "maxValue"  : 143352,
     "minValue"  : 0
 },
 {
     "fieldName" : "total_http",
-    "fieldType" : "int",
+    "fieldType" : "float",
     "maxValue"  : 7045,
     "minValue"  : 638
 },
 {
     "fieldName" : "total_udp",
-    "fieldType" : "int",
+    "fieldType" : "float",
     "maxValue"  : 638,
     "minValue"  : 11
 },
 {
     "fieldName" : "size",
-    "fieldType" : "int",
+    "fieldType" : "float",
     "maxValue"  : 15315931,
     "minValue"  : 13836
 },
 {
     "fieldName" : "size_tcp",
-    "fieldType" : "int",
+    "fieldType" : "float",
     "maxValue"  : 15278447,
     "minValue"  : 0
 },
 {
     "fieldName" : "size_http",
-    "fieldType" : "int",
+    "fieldType" : "float",
     "maxValue"  : 6090358,
     "minValue"  : 0
 },
 {
     "fieldName" : "size_udp",
-    "fieldType" : "int",
+    "fieldType" : "float",
     "maxValue"  : 68596,
     "minValue"  : 1024
 }
